Add utils path to packages for submodules in util module
