+++ demos/navigation-compose/src/androidTest/kotlin/com/bumble/appyx/sample/navigtion/compose/ComposeNavigationRootTest.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.bumble.appyx.navigation.integrationpoint.LocalIntegrationPoint
-import com.bumble.appyx.testing.ui.rules.AppyxTestActivity
+import com.bumble.appyx.utils.testing.ui.rules.AppyxTestActivity
 import org.junit.Rule
 import org.junit.Test
