--- utils/interop-rx3/src/main/kotlin/com/bumble/appyx/interop/rx3/plugin/DisposeOnDestroy.kt
@@ -1,17 +1,0 @@
-package com.bumble.appyx.interop.rx3.plugin
-
-import com.bumble.appyx.interactions.core.plugin.Plugin
-import com.bumble.appyx.navigation.plugin.Destroyable
-import io.reactivex.rxjava3.disposables.CompositeDisposable
-import io.reactivex.rxjava3.disposables.Disposable
-
-private class DisposeOnDestroy(disposables: List<Disposable>) : Destroyable {
-    private val disposable = CompositeDisposable(disposables)
-
-    override fun destroy() {
-        disposable.dispose()
-    }
-}
-
-fun disposeOnDestroyPlugin(vararg disposables: Disposable): Plugin =
-    DisposeOnDestroy(disposables.toList())
