+++ utils/interop-rx3/src/main/kotlin/com/bumble/appyx/utils/interop/rx3/connectable/Connectable.kt
@@ -1,0 +1,9 @@
+package com.bumble.appyx.utils.interop.rx3.connectable
+
+import com.bumble.appyx.navigation.plugin.NodeLifecycleAware
+import com.jakewharton.rxrelay3.Relay
+
+interface Connectable<Input, Output> : NodeLifecycleAware {
+    val input: Relay<Input>
+    val output: Relay<Output>
+}
