+++ utils/testing-junit4/src/main/kotlin/com/bumble/appyx/utils/testing/junit4/MainDispatcherRule.kt
@@ -1,0 +1,21 @@
+package com.bumble.appyx.utils.testing.junit4
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import org.junit.rules.TestWatcher
+import org.junit.runner.Description
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class MainDispatcherRule : TestWatcher() {
+
+    override fun starting(description: Description) {
+        Dispatchers.setMain(Dispatchers.Unconfined)
+    }
+
+    override fun finished(description: Description) {
+        Dispatchers.resetMain()
+    }
+
+}
