--- utils/testing-junit5/src/main/kotlin/com/bumble/appyx/testing/junit5/util/CoroutinesTestExtension.kt
@@ -1,24 +1,0 @@
-package com.bumble.appyx.testing.junit5.util
-
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.TestDispatcher
-import kotlinx.coroutines.test.UnconfinedTestDispatcher
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
-import org.junit.jupiter.api.extension.AfterEachCallback
-import org.junit.jupiter.api.extension.BeforeEachCallback
-import org.junit.jupiter.api.extension.ExtensionContext
-
-@OptIn(ExperimentalCoroutinesApi::class)
-class CoroutinesTestExtension(
-    private val dispatcher: TestDispatcher = UnconfinedTestDispatcher(),
-) : BeforeEachCallback, AfterEachCallback {
-    override fun beforeEach(context: ExtensionContext?) {
-        Dispatchers.setMain(dispatcher)
-    }
-
-    override fun afterEach(context: ExtensionContext?) {
-        Dispatchers.resetMain()
-    }
-}
