--- utils/testing-junit5/src/main/kotlin/com/bumble/appyx/testing/junit5/util/InstantExecutorExtension.kt
@@ -1,27 +1,0 @@
-package com.bumble.appyx.testing.junit5.util
-
-import android.annotation.SuppressLint
-import androidx.arch.core.executor.ArchTaskExecutor
-import androidx.arch.core.executor.TaskExecutor
-import org.junit.jupiter.api.extension.AfterEachCallback
-import org.junit.jupiter.api.extension.BeforeEachCallback
-import org.junit.jupiter.api.extension.ExtensionContext
-
-class InstantExecutorExtension : BeforeEachCallback, AfterEachCallback {
-
-    @SuppressLint("RestrictedApi")
-    override fun beforeEach(context: ExtensionContext?) {
-        ArchTaskExecutor.getInstance()
-            .setDelegate(object : TaskExecutor() {
-                override fun executeOnDiskIO(runnable: Runnable) = runnable.run()
-                override fun postToMainThread(runnable: Runnable) = runnable.run()
-                override fun isMainThread(): Boolean = true
-            })
-    }
-
-    @SuppressLint("RestrictedApi")
-    override fun afterEach(context: ExtensionContext?) {
-        ArchTaskExecutor.getInstance().setDelegate(null)
-    }
-
-}
